semplifica la funzione di creazione dei punti del arco

+++ include/odometry.h
@@ -11,25 +11,25 @@
 extern float mm_sx;
 extern float mm_dx;
 
-extern double pose[3][3];
-extern double newPose[3][3];
-extern double rt[3][3];
-extern double r[3][3];
-extern double t1[3][3];
-extern double t2[3][3];
-extern double temp[3][3];
+extern float pose[3][3];
+extern float newPose[3][3];
+extern float rt[3][3];
+extern float r[3][3];
+extern float t1[3][3];
+extern float t2[3][3];
+extern float temp[3][3];
 
 typedef struct {
     pthread_mutex_t lock;
-    double x;
-    double y;
-    double theta;
+    float x;
+    float y;
+    float theta;
 } position_t;
 
 extern position_t position;
 
 
-void moltiplica_matrici_3x3(double matA[3][3], double matB[3][3], double result[3][3]);
+void moltiplica_matrici_3x3(float matA[3][3], float matB[3][3], float result[3][3]);
 void findNewPose(int l_ticks_odo, int r_ticks_odo);
 
 #endif // ODOMETRY_H
